Moved python generator to separate subpackage
