+++ simphony_metatools/owl/owl_json_ld_generator.py
@@ -1,0 +1,12 @@
+import json
+
+
+ONTOLOGY_PREFIX = "http://simphony#"
+OWL_PREFIX = "http://www.w3.org/2002/07/owl#"
+RDF_SCHEMA_PREFIX = "http://www.w3.org/2000/01/rdf-schema#"
+
+
+class OWLJSONLDGenerator():
+    def generate(self, ontology, output_path, overwrite=False):
+        pass
+
