+++ simphony_metatools/python/meta_class_generator.py
@@ -3,9 +3,8 @@ from __future__ import print_function
 import os
 
 from simphony_metaparser.utils import traverse
-
-from .single_meta_class_generator import SingleMetaClassGenerator
-from . import utils
+from simphony_metatools import utils
+from simphony_metatools.python.single_meta_class_generator import SingleMetaClassGenerator
 
 
 class MetaClassGenerator(object):
